requirements.txt
sqlalchemy
werkzeug
flask
python-dotenv
flask-login
flask-sqlalchemy
flask-limiter
flask-wtf
